Simplify logic for the 'init' variable related to minutes: - rename to 'init_min' and make it an uint8_t - cycle was 3 -> 2 -> 0 (not 3 -> 1 -> 0), turn this into 2 -> 1 -> 0 - simplify state switching in mainloop()

+++ mainloop.h
@@ -71,7 +71,7 @@ int mainloop(char *logfilename,
     void (*display_weather)(void),
     void (*display_time)(uint32_t, struct tm),
     void (*display_thirdparty_buffer)(uint8_t *),
-    void (*set_time)(int, uint32_t, uint16_t, int, struct tm),
+    void (*set_time)(uint8_t, uint32_t, uint16_t, int, struct tm),
     void (*process_input)(uint16_t *, int, char *, int *, int *),
     void (*post_process_input)(char **, int *, uint16_t *, int));
